feat: 资产查询修改

+++ asset-vo/src/main/java/com/antiy/asset/vo/response/AssetSoftwareInstallResponse.java
@@ -85,7 +85,7 @@ public class AssetSoftwareInstallResponse {
      * 安装时间
      */
     @ApiModelProperty("安装时间")
-    private Integer installTime;
+    private Long installTime;
 
     public Integer getUserId() {
         return userId;
@@ -175,11 +175,11 @@ public class AssetSoftwareInstallResponse {
         this.installStatus = installStatus;
     }
 
-    public Integer getInstallTime() {
+    public Long getInstallTime() {
         return installTime;
     }
 
-    public void setInstallTime(Integer installTime) {
+    public void setInstallTime(Long installTime) {
         this.installTime = installTime;
     }
